requirements.txt
# Core dependencies
crewai>=0.148.0
python-telegram-bot>=22.3
python-dotenv>=1.1.1
aiofiles>=24.1.0

# LLM and AI
google-generativeai>=0.3.0
litellm>=1.72.6

# MCP Server
mcp>=1.0.0

# HTTP and API
httpx>=0.28.1
requests>=2.32.4

# Data processing
pydantic>=2.4.2
jsonschema>=4.25.0
pyyaml>=6.0.0

# Security
bcrypt>=4.0.1
cryptography>=45.0.5

# Utilities
click>=8.1.7
rich>=14.0.0
tqdm>=4.65.0